initial work on attachments
